workspace refactor, jest testing with coverage --bloated
