+++ .devcontainer/devcontainer.json
@@ -4,6 +4,9 @@
 	"name": "botOrlin App",
 	"dockerComposeFile": "docker-compose.yml",
 	"service": "frontend",
+	// "build": {
+	// 	"dockerfile": "../frontend/Dockerfile",
+	// },
 	"workspaceFolder": "/workspace",
 	// Set *default* container specific settings.json values on container create.
 	// "settings": { 
@@ -21,7 +24,9 @@
 		3000
 	],
 	// Use 'postCreateCommand' to run commands after the container is created.
-	// "postCreateCommand": "pip3 install --user -r requirements.txt",
+	// "postCreateCommand": "pip3 install --user -r requirements.txt",¨
+	"postCreateCommand": "cd /workspace/frontend && yarn install",
+	"postStartCommand": "cd /workspace/frontend && yarn dev",
 	// Comment out connect as root instead. More info: https://aka.ms/vscode-remote/containers/non-root.
 	"remoteUser": "node"
 }
