+++ frontend/Dockerfile
@@ -1,7 +1,7 @@
 # See here for image contents: https://github.com/microsoft/vscode-dev-containers/tree/v0.195.0/containers/javascript-node/.devcontainer/base.Dockerfile
 # [Choice] Node.js version (use -bullseye variants on local arm64/Apple Silicon): 16, 14, 12, 16-bullseye, 14-bullseye, 12-bullseye, 16-buster, 14-buster, 12-buster
 ARG VARIANT=16
-FROM mcr.microsoft.com/vscode/devcontainers/javascript-node:0-${VARIANT}
+FROM mcr.microsoft.com/vscode/devcontainers/javascript-node:16
 
 # # Create app directory
 RUN mkdir -p /workspace/frontend
@@ -10,7 +10,7 @@ WORKDIR /workspace/frontend
 # Installing dependencies
 COPY package*.json ./
 
-RUN yarn install
+# RUN yarn install
 
-# Running the app dev server
-CMD ["yarn", "dev"]
+# # Running the app dev server
+# CMD ["yarn", "dev"]
