+++ frontend/src/lib/getHello.ts
