+++ frontend/src/lib/myMath.ts
