+++ frontend/src/styles/globals.css
