--- tests/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "tests",
-  "version": "0.1.0",
-  "description": "Tests for botOrlin App",
-  "author": {
-    "name": "Daniel Kňourek",
-    "email": "dev.danielknourek@gmail.com"
-  },
-  "scripts": {
-    "test": "jest",
-    "testc": "jest --coverage"
-  },
-  "license": "MIT",
-  "private": true,
-  "devDependencies": {
-    "@babel/core": "^7.17.10",
-    "@babel/preset-env": "^7.17.10",
-    "@babel/preset-typescript": "^7.16.7",
-    "@jest/types": "^28.1.0",
-    "@testing-library/dom": "^8.13.0",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.2.0",
-    "@types/jest": "^27.5.0",
-    "@types/node": "^17.0.31",
-    "babel-jest": "^28.1.0",
-    "frontend": "file:./../frontend",
-    "jest": "^28.1.0",
-    "jest-cli": "^28.1.0",
-    "jest-css-modules": "^2.1.0",
-    "jest-css-modules-transform": "^4.4.2",
-    "ts-jest": "^28.0.1",
-    "ts-node": "^10.7.0",
-    "typescript": "^4.6.4"
-  }
-}
