--- tests/tsconfig.json
@@ -1,45 +1,0 @@
-{
-  "compilerOptions": {
-    // settings from nextjs template
-    "target": "es5",
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ],
-    "allowJs": true,
-    "skipLibCheck": true,
-    "strict": true,
-    "forceConsistentCasingInFileNames": true,
-    "noEmit": true,
-    "esModuleInterop": true,
-    "module": "esnext",
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
-    "isolatedModules": true,
-    "jsx": "preserve",
-    "incremental": true,
-    // extra from https://github.com/Swizec/jest-db-mock-example/blob/main/tsconfig.json
-    // "target": "es2017",
-    "types": [
-      "node",
-      "jest"
-    ],
-    "allowSyntheticDefaultImports": true,
-    // "module": "commonjs",
-    "baseUrl": ".",
-    "noFallthroughCasesInSwitch": true,
-    "outDir": "./dist"
-  },
-  "compileOnSave": true,
-  "include": [
-    "next-env.d.ts",
-    "**/*.ts",
-    "**/*.tsx",
-    "tsconfig.test.json",
-    "jest.config.ts"
-  ],
-  "exclude": [
-    "node_modules/(?!frontend)"
-  ]
-}
